refactor my anime list functions
